fix(web): split workspace layout into server+client for Next.js 16 compatibility
- WorkspaceScopeLayout is now a Server Component that calls connection()
  once, opting the entire [id] subtree out of static prerendering
- Extracted Zustand/useParams gating logic into WorkspaceScopeClient
- Wraps HydrationProvider in <Suspense> in root layout — required by
  cacheComponents when a client component reads request-time data (session)
- All workspace/* pages now render as Partial Prerender (◐) correctly
- Removed incorrect dynamic exports and unstable_noStore calls from pages
