+++ apps/marketing/next.config.ts
@@ -3,9 +3,7 @@ import type { NextConfig } from "next";
 const nextConfig: NextConfig = {
   output: "export",
   transpilePackages: ["@context-layer/ui"],
-  eslint: {
-    ignoreDuringBuilds: true,
-  },
+  reactCompiler: true,
 };
 
 export default nextConfig;
