+++ apps/marketing/package.json
@@ -14,17 +14,21 @@
     "@context-layer/ui": "workspace:*",
     "lucide-react": "^1.8.0",
     "motion": "^12.38.0",
-    "next": "15.5.15",
-    "react": "19.1.0",
-    "react-dom": "19.1.0"
+    "next": "16.2.4",
+    "react": "19.2.5",
+    "react-dom": "19.2.5"
   },
   "devDependencies": {
     "@playwright/test": "^1.59.1",
     "@tailwindcss/postcss": "^4",
     "@types/node": "^20",
-    "@types/react": "^19",
-    "@types/react-dom": "^19",
+    "@types/react": "19.2.14",
+    "@types/react-dom": "19.2.3",
     "tailwindcss": "^4",
     "typescript": "^5"
+  },
+  "overrides": {
+    "@types/react": "19.2.14",
+    "@types/react-dom": "19.2.3"
   }
 }
