+++ apps/web/src/app/(playground)/workspace/[id]/chatbot/page.tsx
@@ -1,11 +1,8 @@
-"use client";
-
-import { useParams } from "next/navigation";
 import { FullChatbot } from "@/components/playground/chatbot/full-chatbot";
+import { connection } from "next/server";
 
-export default function ChatbotPage() {
-  const params = useParams();
-  const id = Array.isArray(params?.id) ? params.id[0] : (params?.id as string | undefined);
-  if (!id) return null;
+export default async function ChatbotPage({ params }: { params: Promise<{ id: string }> }) {
+  await connection();
+  const { id } = await params;
   return <FullChatbot workspaceId={id} />;
 }
