+++ apps/web/src/app/(playground)/workspace/[id]/generate/docsgen/[bundle]/page.tsx
@@ -1,18 +1,14 @@
-"use client";
-
-import { notFound, useParams } from "next/navigation";
+import { notFound } from "next/navigation";
 import { DocsGenBundleView } from "@/components/playground/generate/docsgen/bundle-view";
 import { isValidBundleSlug } from "@/components/playground/generate/docsgen/catalog";
 
-export default function DocsGenBundlePage() {
-  const params = useParams();
-  const id = Array.isArray(params?.id) ? params.id[0] : (params?.id as string | undefined);
-  const bundleRaw = Array.isArray(params?.bundle)
-    ? params.bundle[0]
-    : (params?.bundle as string | undefined);
+export default async function DocsGenBundlePage({
+  params,
+}: { params: Promise<{ id: string; bundle: string }> }) {
+  const { id, bundle } = await params;
 
-  if (!id || !bundleRaw) return null;
-  if (!isValidBundleSlug(bundleRaw)) notFound();
+  if (!id || !bundle) return null;
+  if (!isValidBundleSlug(bundle)) notFound();
 
-  return <DocsGenBundleView workspaceId={id} bundle={bundleRaw} />;
+  return <DocsGenBundleView workspaceId={id} bundle={bundle} />;
 }
