+++ apps/web/src/app/(playground)/workspace/[id]/generate/docsgen/page.tsx
@@ -1,6 +1,6 @@
 import { redirect } from "next/navigation";
 
-export default async function DocsGenPage(props: { params: Promise<{ id: string }> }) {
-  const { id } = await props.params;
+export default async function DocsGenPage({ params }: { params: Promise<{ id: string }> }) {
+  const { id } = await params;
   redirect(`/workspace/${id}/generate/docsgen/structure-architecture`);
 }
