+++ apps/web/src/app/(playground)/workspace/[id]/generate/omniboard/page.tsx
@@ -1,11 +1,6 @@
-"use client";
-
-import { useParams } from "next/navigation";
 import { OmniBoardSurface } from "@/components/playground/generate/omniboard/surface";
 
-export default function OmniBoardPage() {
-  const params = useParams();
-  const id = Array.isArray(params?.id) ? params.id[0] : (params?.id as string | undefined);
-  if (!id) return null;
+export default async function OmniBoardPage({ params }: { params: Promise<{ id: string }> }) {
+  const { id } = await params;
   return <OmniBoardSurface workspaceId={id} />;
 }
