+++ apps/web/src/app/(playground)/workspace/[id]/intelligence/dependencies/page.tsx
@@ -1,13 +1,8 @@
-"use client";
-
-import { useParams } from "next/navigation";
 import { IntelligencePageShell } from "@/components/playground/intelligence/page-shell";
 import { IntelligenceDependencies } from "@/components/playground/intelligence/dependencies/dashboard";
 
-export default function DependenciesPage() {
-  const params = useParams();
-  const id = Array.isArray(params?.id) ? params.id[0] : (params?.id as string | undefined);
-  if (!id) return null;
+export default async function DependenciesPage({ params }: { params: Promise<{ id: string }> }) {
+  const { id } = await params;
   return (
     <IntelligencePageShell workspaceId={id}>
       <IntelligenceDependencies workspaceId={id} />
