+++ apps/web/src/app/(playground)/workspace/[id]/wiki/logs/page.tsx
@@ -1,15 +1,8 @@
-"use client";
-
-import { useParams } from "next/navigation";
-import { useStore } from "@/stores";
-import { EmptyStateNeedsWiki } from "@/components/playground/wiki/empty-state-needs-wiki";
 import { WikiLogsView } from "@/components/playground/wiki/logs/view";
+import { connection } from "next/server";
 
-export default function WikiLogsPage() {
-  const params = useParams();
-  const id = Array.isArray(params?.id) ? params.id[0] : (params?.id as string | undefined);
-  const hasWiki = useStore((s) => s.workspaces.find((w) => w.id === id)?.hasWiki ?? false);
-  if (!id) return null;
-  if (!hasWiki) return <EmptyStateNeedsWiki workspaceId={id} tabLabel="Logs" />;
+export default async function WikiLogsPage({ params }: { params: Promise<{ id: string }> }) {
+  await connection();
+  const { id } = await params;
   return <WikiLogsView workspaceId={id} />;
 }
