+++ apps/web/src/app/(playground)/workspace/[id]/wiki/view/page.tsx
@@ -1,17 +1,8 @@
-"use client";
-
-import { useParams } from "next/navigation";
-import { useStore } from "@/stores";
-import { EmptyStateNeedsWiki } from "@/components/playground/wiki/empty-state-needs-wiki";
 import { WikiViewer } from "@/components/playground/wiki/view/viewer";
+import { connection } from "next/server";
 
-export default function WikiViewPage() {
-  const params = useParams();
-  const id = Array.isArray(params?.id) ? params.id[0] : (params?.id as string | undefined);
-  const isHydrated = useStore((s) => s.isHydrated);
-  const hasWiki = useStore((s) => s.workspaces.find((w) => w.id === id)?.hasWiki ?? false);
-  if (!id) return null;
-  if (!isHydrated) return null;
-  if (!hasWiki) return <EmptyStateNeedsWiki workspaceId={id} tabLabel="View" />;
+export default async function WikiViewPage({ params }: { params: Promise<{ id: string }> }) {
+  await connection();
+  const { id } = await params;
   return <WikiViewer workspaceId={id} />;
 }
