+++ apps/web/tsconfig.json
@@ -1,7 +1,11 @@
 {
   "compilerOptions": {
     "target": "ES2017",
-    "lib": ["dom", "dom.iterable", "esnext"],
+    "lib": [
+      "dom",
+      "dom.iterable",
+      "esnext"
+    ],
     "allowJs": true,
     "skipLibCheck": true,
     "strict": true,
@@ -11,7 +15,7 @@
     "moduleResolution": "bundler",
     "resolveJsonModule": true,
     "isolatedModules": true,
-    "jsx": "preserve",
+    "jsx": "react-jsx",
     "incremental": true,
     "plugins": [
       {
@@ -19,15 +23,37 @@
       }
     ],
     "paths": {
-      "@/*": ["./src/*"],
-      "@context-layer/ui/*": ["../../packages/ui/src/*"],
-      "components/*": ["../../packages/ui/src/components/*"],
-      "utils": ["../../packages/ui/src/lib/utils"],
-      "ui/*": ["../../packages/ui/src/components/ui/*"],
-      "lib/*": ["../../packages/ui/src/lib/*"],
-      "hooks/*": ["../../packages/ui/src/hooks/*"]
+      "@/*": [
+        "./src/*"
+      ],
+      "@context-layer/ui/*": [
+        "../../packages/ui/src/*"
+      ],
+      "components/*": [
+        "../../packages/ui/src/components/*"
+      ],
+      "utils": [
+        "../../packages/ui/src/lib/utils"
+      ],
+      "ui/*": [
+        "../../packages/ui/src/components/ui/*"
+      ],
+      "lib/*": [
+        "../../packages/ui/src/lib/*"
+      ],
+      "hooks/*": [
+        "../../packages/ui/src/hooks/*"
+      ]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
-  "exclude": ["node_modules"]
+  "include": [
+    "next-env.d.ts",
+    "**/*.ts",
+    "**/*.tsx",
+    ".next/types/**/*.ts",
+    ".next/dev/types/**/*.ts"
+  ],
+  "exclude": [
+    "node_modules"
+  ]
 }
